Sign error

+++ src/neuralnetwork/neuralnetwork.cpp
@@ -1279,7 +1279,7 @@ void NeuralNetwork::log_training_info(
         tab, tab, tab, tab, tab, tab, "direction          : ", details.get_error_evaluation_config().direction_lambda(), "\n",
         tab, tab, tab, tab, tab, tab, "cross-entropy      : ", details.get_error_evaluation_config().cross_entropy_lambda(), "\n",
         tab, tab, tab, tab, "use direction penalty: ", details.get_error_evaluation_config().use_direction_penalty() ? "true" : "false");
-      if (output_layer_index < output_layer_details.size())
+      if (output_layer_index < static_cast<int>(output_layer_details.size()))
       {
         output_layer_details_string += "\n";
       }
